Extract common constants to file
