+++ packages/edit-site/src/components/dataviews/constants.js
@@ -1,0 +1,5 @@
+// Field types.
+export const ENUMERATION_TYPE = 'enumeration';
+
+// Filter operators.
+export const OPERATOR_IN = 'in';
