+++ packages/edit-site/src/components/dataviews/in-filter.js
@@ -7,8 +7,10 @@ import {
 } from '@wordpress/components';
 import { __, sprintf } from '@wordpress/i18n';
 
-// TODO: find a place where these constants can be shared across components.
-export const OPERATOR_IN = 'in';
+/**
+ * Internal dependencies
+ */
+import { OPERATOR_IN } from './constants';
 
 export default ( { filter, view, onChangeView } ) => {
 	const valueFound = view.filters.find(
